Add footer to topics board pages
When you're in a room, you might want to edit its details or get back to
the home page without manually editing the URL. Now you can! Just scroll
down.

I really liked that the page by default only shows the board, and that
as soon as you start to scroll down, you see the footer. I copied this
on the home page, to make the list of rooms take up the entire view
height, as well. And since people who already use the app don't need to
see the intro stuff, I made the main links back to the homepage go to
just that bottom section, by default.

Getting the text in the footer to be vertically centered was totally
crazy-bonkers-crazy! Wow. CSS!! Even with flex-box, it's still totally
whacked sometimes. Although flexbox certainly made it a lot better!
After learning some flexbox tricks, I went and made the grid of rooms on
the homepage look nicer, too.

The CSS feels sorta unorganized to me. I've never worked on a team where
it actually feels organized. I feel like this project is reaching a
point where a better organizational scheme would be really useful. But I
don't know if it's worth learning how to do that. I feel like knowing
about how to organize CSS files is a designer's job. Maybe that's lazy
of me. :-/